d/rules: Install ethtool in /usr/sbin
The backport-safe way to do this is to use dh_movetousr which will be
a no-op in bookworm-backports (but isn't currently available there).

However, ethtool isn't in any backports suites currently, so make the
simpler change of deleting the sbindir=/sbin parameter to configure.
Add a comment for the benefit of anyone to cares to backport ethtool
later.